refactor: email package명 external로 변경
